[svn r286] Type of style defined in enumeration

+++ src/src/com/panayotis/jubler/subs/loader/text/format/StyledFormat.java
@@ -31,18 +31,18 @@ import com.panayotis.jubler.subs.style.StyleType;
  * @author teras
  */
 public class StyledFormat {
-    public final static int COLOR_NORMAL = 0;
-    public final static int COLOR_REVERSE = 1;
-    public final static int COLOR_ALPHA_NORMAL = 2;
-    public final static int COLOR_ALPHA_REVERSE = 3;
+    public final static byte COLOR_NORMAL = 0;
+    public final static byte COLOR_REVERSE = 1;
+    public final static byte COLOR_ALPHA_NORMAL = 2;
+    public final static byte COLOR_ALPHA_REVERSE = 3;
     
-    public final static int FORMAT_UNDEFINED = 0;
-    public final static int FORMAT_STRING = 1;
-    public final static int FORMAT_INTEGRAL = 2;
-    public final static int FORMAT_REAL = 3;
-    public final static int FORMAT_FLAG = 4;
-    public final static int FORMAT_COLOR = 5;
-    public final static int FORMAT_DIRECTION = 6;
+    public final static byte FORMAT_UNDEFINED = 0;
+    public final static byte FORMAT_STRING = 1;
+    public final static byte FORMAT_INTEGRAL = 2;
+    public final static byte FORMAT_REAL = 3;
+    public final static byte FORMAT_FLAG = 4;
+    public final static byte FORMAT_COLOR = 5;
+    public final static byte FORMAT_DIRECTION = 6;
     
     public StyleType style;
     public String tag;
